G-9700 refactor
